requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
Pillow
tqdm
streamlit-webrtc
av
